+++ pi4j-core/src/main/java/com/pi4j/io/gpio/impl/GpioEventMonitorImpl.java
@@ -76,7 +76,7 @@ public class GpioEventMonitorImpl implements PinListener
             }
             else if(event.getEventType() == PinEventType.ANALOG_VALUE_CHANGE)
             {
-                int value = ((PinAnalogValueChangeEvent)event).getValue();
+                double value = ((PinAnalogValueChangeEvent)event).getValue();
 
                 // process event callbacks for analog listeners
                 for (GpioPinListener listener : pin.getListeners())
